requirements.txt
haystack-ai
pinecone-haystack
sentence-transformers
nltk>=3.9.1
fastapi
uvicorn
python-dotenv
pypdf
google-generativeai
-e .
